backend/requirements.txt
flask
tensorflow
pillow
matplotlib
numpy
reportlab
